fix(pack): remove `opts.force` from `vim.pack.update()`
The decision about adding this is postponed until there is lockfile
support. The reason is that lockfile will enable the alternative
workflow of `add()` ensuring that plugin on disk matches `version`.
This will remove the need for calling `vim.pack.update()` after changing
`version` at the cost of not being able to preview and confirm/deny
changes to be done to a plugin on disk. They will still be present in
the log file, though.

The only downside right now is that changing `version` of already
installed plugin requires Internet connection. Which can be
reasonable enough as a temporary (hopefully not for long) restriction.